Make the docker scripts actually work, sorta.
start_machine.bat calls docker-machine start, then gets the env via docker-machine env and saves it in the bs-meter-dev.bat file, which it then calls immediately to source the env into the shell. Then it calls docker-machine ip to get the IP of the new machine and stores it into bs-meter-dev-address.txt.